+++ src/config/defaultConfig.ts
@@ -11,7 +11,7 @@ export const defaultConfig: Config = {
       width: 1280,
       height: 720
     },
-    userAgent: "wqg/3.2.6",
+    userAgent: "wqg/3.2.7",
     locale: "en-US",
     colorScheme: "light"
   },